+++ lib/kb_Msuite/kb_MsuiteClient.py
@@ -47,16 +47,17 @@ class kb_Msuite(object):
            the tetra command) dist_value - when running dist_plot, set this
            to a value between 0 and 100 thread -  number of threads
            reduced_tree - if set to 1, run checkM with the reduced_tree flag,
-           which will keep memory limited to less than 16gb quiet - pass the
-           --quite parameter to checkM, but doesn't seem to work for all
-           subcommands) -> structure: parameter "subcommand" of String,
-           parameter "bin_folder" of String, parameter "out_folder" of
-           String, parameter "plots_folder" of String, parameter "seq_file"
-           of String, parameter "tetra_file" of String, parameter
-           "dist_value" of Long, parameter "thread" of Long, parameter
-           "reduced_tree" of type "boolean" (A boolean - 0 for false, 1 for
-           true. @range (0, 1)), parameter "quiet" of type "boolean" (A
-           boolean - 0 for false, 1 for true. @range (0, 1))
+           which will keep memory limited to less than 16gb (otherwise needs
+           40+ GB, which NJS worker nodes do have) quiet - pass the --quite
+           parameter to checkM, but doesn't seem to work for all subcommands)
+           -> structure: parameter "subcommand" of String, parameter
+           "bin_folder" of String, parameter "out_folder" of String,
+           parameter "plots_folder" of String, parameter "seq_file" of
+           String, parameter "tetra_file" of String, parameter "dist_value"
+           of Long, parameter "thread" of Long, parameter "reduced_tree" of
+           type "boolean" (A boolean - 0 for false, 1 for true. @range (0,
+           1)), parameter "quiet" of type "boolean" (A boolean - 0 for false,
+           1 for true. @range (0, 1))
         """
         return self._client.call_method(
             'kb_Msuite.run_checkM',
@@ -65,13 +66,14 @@ class kb_Msuite(object):
     def run_checkM_lineage_wf(self, params, context=None):
         """
         :param params: instance of type "CheckMLineageWfParams" (input_ref -
-           reference to the input Assembly or BinnedContigs data (could be
-           expanded to include Genome objects as well)) -> structure:
-           parameter "input_ref" of String, parameter "workspace_name" of
-           String, parameter "save_output_dir" of type "boolean" (A boolean -
-           0 for false, 1 for true. @range (0, 1)), parameter
-           "save_plots_dir" of type "boolean" (A boolean - 0 for false, 1 for
-           true. @range (0, 1))
+           reference to the input Assembly, AssemblySet, Genome, GenomeSet,
+           or BinnedContigs data) -> structure: parameter "input_ref" of
+           String, parameter "workspace_name" of String, parameter
+           "reduced_tree" of type "boolean" (A boolean - 0 for false, 1 for
+           true. @range (0, 1)), parameter "save_output_dir" of type
+           "boolean" (A boolean - 0 for false, 1 for true. @range (0, 1)),
+           parameter "save_plots_dir" of type "boolean" (A boolean - 0 for
+           false, 1 for true. @range (0, 1))
         :returns: instance of type "CheckMLineageWfResult" -> structure:
            parameter "report_name" of String, parameter "report_ref" of String
         """