+++ RELEASE_NOTES.md
@@ -1,7 +1,15 @@
+### Version 1.2.1
+__Changes__
+- Added choice to use reduced or full Reference Tree to input widget
+- Changed save output and save plots input widgets from checkbox to dropdown
+- Added AssemblySet, Genome, and GenomeSet input types
+
 ### Version 1.2.0
+__Changes__
 - Updated CheckM to v1.0.8 which fixes failure with --reduced_tree for certain ribosomal proteins
 
 ### Version 1.1.2
+__Changes__
 - Fix infinite recursion in output folder compression
 - Fix running on empty FASTA